+++ R/globals.R
@@ -8,6 +8,7 @@ utils::globalVariables(c(
   ".data", # <as_epidist_linelist_data.epidist_aggregate_data>
   ".data", # <.prepare_marginal_data>
   ".data", # <as_epidist_naive_model.epidist_linelist_data>
+  ".data", # <plot_events>
   "rlnorm", # <simulate_secondary>
   ".data", # <simulate_secondary>
   ".data", # <simulate_dates>